Refactor DocumentStructureToolWindow
