+++ packages/main/src/react-app/toolwindow/DocumentStructureToolWindow/TagsPanel.module.css
